adding custom error and creating utilities folder
