+++ src/controllers/user_controller.js
@@ -6,9 +6,9 @@ const create_user = async (req, res) => {
     try {
         const creat_user = await userServices.creat_user(body);
         return res.status(201).send({ message: "Usuário criado com sucesso", creat_user });
-    } catch (error) { // Adicionado o parâmetro error
-        console.error('Erro ao criar usuário:', error);
-        return res.status(500).send({ message: "Erro ao criar usuário", error: error.message });
+    } catch (err) { // Adicionado o parâmetro error
+        const statusCode = err.statusCode || 500;
+        return res.status(statusCode).send({ message: err.message });
     }
 }
