+++ src/services/user_services.js
@@ -3,33 +3,8 @@ import dioceseModels from '../models/diocese_models.js';
 import globalmiddleware from '../middlwares/global_middlewares.js';
 import bcrypt from 'bcrypt';
 import logger from '../logger/logger.config.js';
-
-
-const generateRegistrationId = async (diocese_id) => {
-    logger.info('Generating registration_id');
-    try {
-
-        const randomPart = Math.floor(1000 + Math.random() * 9000).toString().trim();
-        const dioceseIdPart = diocese_id.diocese_id.toString().trim();
-        const registration_id = parseInt(`${randomPart}${dioceseIdPart}`);
-
-        const existingUser = await userModels.find_user_by_registration_id(registration_id);
-        if (existingUser) {
-            return generateRegistrationId(diocese_id);
-        }
-
-        return registration_id;
-
-    }catch (error) {
-        logger.error('Erro ao gerar registration_id:', error);
-        throw new Error('Erro ao gerar registration_id');
-    }
-}
-
-const formatDateForDatabase = (dateString) => {
-    const [day, month, year] = dateString.split('/');
-    return `${year}-${month}-${day}`; // Retorna no formato yyyy-mm-dd
-};
+import CustomError from '../utils/CustomError.js';
+import { generateRegistrationId, formatDateForDatabase } from '../utils/basicFunctions.js';
 
 const creat_user = async (body) => {
     logger.info('Creating user');
@@ -37,32 +12,32 @@ const creat_user = async (body) => {
 
     if (!name || !email || !password || !phone || !birth_date || !name_diocese) {
         logger.error('All fields are required');
-        throw new Error('All fields are required');
+        throw new CustomError('All fields are required', 400);
     }
 
     const verify_email_registration = await userModels.find_user_by_email(email);
     if (verify_email_registration) {
         logger.error('Email already registered');
-        throw new Error('Email already registered');
+        throw new CustomError('Email already registered', 409);
     }
 
     const diocese_id = await dioceseModels.find_diocese_by_name(name_diocese);
     if (!diocese_id) {
         logger.error('Diocese not found');
-        throw new Error('Diocese not found');
+        throw new CustomError('Diocese not found', 400);
     }
 
     const registration_id = await generateRegistrationId(diocese_id);
     if (!registration_id) {
         logger.error('Error generating registration ID');
-        throw new Error('Error generating registration ID');
+        throw new CustomError('Error generating registration ID');
     }
 
     const birth_formated = formatDateForDatabase(birth_date);
 
     if (password.length > 20 || password.length < 6) {
         logger.error('Password must be between 6 and 20 characters');
-        throw new Error('Password must be between 6 and 20 characters');
+        throw new Error('Password must be between 6 and 20 characters', 400);
     }
 
     const hashed_password = await bcrypt.hash(password, 10);
@@ -70,13 +45,13 @@ const creat_user = async (body) => {
     const user_rcc = await userModels.create_user(registration_id, name, email, hashed_password, phone, birth_formated, diocese_id.diocese_id);
     if (!user_rcc) {
         logger.error('Error creating user or missing required fields user');
-        throw new Error('Error Error creating user or missing required fields user');
+        throw new CustomError('Error Error creating user or missing required fields user', 400);
     }
 
     const generateToken = await globalmiddleware.generateToken(user_rcc.registration_id, user_rcc.niveluser);
     if (!generateToken) {
         logger.error('Error generating token');
-        throw new Error('Error generating token');
+        throw new CustomError('Error generating token', 400);
     }
 
     logger.info('User created successfully');
