+++ src/utils/CustomError.js
@@ -1,0 +1,8 @@
+class CustomError extends Error {
+    constructor(message, statusCode) {
+        super(message);
+        this.statusCode = statusCode;
+    }
+}
+
+export default CustomError;
